@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 0.1.0
 
 environment:
-  sdk: '>=3.6.0 <4.0.0'
+  sdk: '>=3.7.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -15,6 +15,7 @@ dependencies:
   audioplayers: ^6.3.0
   path_provider: ^2.1.5
   path: ^1.9.1
+  shared_preferences: ^2.2.2
 
 dev_dependencies:
   flutter_test:
