@@ -15,10 +15,10 @@ class Avatar {
     String? color,
     List<Voice>? voices,
     IconData? icon,
-  })  : id = id ?? const Uuid().v4(),
-        color = color ?? _getRandomColor(),
-        voices = voices ?? [],
-        icon = icon ?? _getDefaultIcon();
+  }) : id = id ?? const Uuid().v4(),
+       color = color ?? _getRandomColor(),
+       voices = voices ?? [],
+       icon = icon ?? _getDefaultIcon();
 
   /// Creates a copy of this Avatar with the given fields replaced with the new values
   Avatar copyWith({
@@ -38,14 +38,7 @@ class Avatar {
 
   /// Generate a random color from a predefined set of colors
   static String _getRandomColor() {
-    const colors = [
-      'blue',
-      'purple',
-      'pink',
-      'orange',
-      'green',
-      'teal',
-    ];
+    const colors = ['blue', 'purple', 'pink', 'orange', 'green', 'teal'];
     return colors[DateTime.now().microsecond % colors.length];
   }
 
@@ -62,6 +55,10 @@ class Voice {
   final String audioUrl;
   final Duration duration;
   final DateTime createdAt;
+  final String category;
+  final int playCount;
+  final DateTime? lastPlayed;
+  final String? color; // Color of the voice item
 
   Voice({
     String? id,
@@ -69,14 +66,22 @@ class Voice {
     required this.audioUrl,
     Duration? duration,
     DateTime? createdAt,
-  })  : id = id ?? const Uuid().v4(),
-        duration = duration ?? Duration.zero,
-        createdAt = createdAt ?? DateTime.now();
+    this.category = 'Uncategorized',
+    this.playCount = 0,
+    this.lastPlayed,
+    this.color,
+  }) : id = id ?? const Uuid().v4(),
+       duration = duration ?? Duration.zero,
+       createdAt = createdAt ?? DateTime.now();
 
   Voice copyWith({
     String? name,
     String? audioUrl,
     Duration? duration,
+    String? category,
+    int? playCount,
+    DateTime? lastPlayed,
+    String? color,
   }) {
     return Voice(
       id: id,
@@ -84,6 +89,14 @@ class Voice {
       audioUrl: audioUrl ?? this.audioUrl,
       duration: duration ?? this.duration,
       createdAt: createdAt,
+      category: category ?? this.category,
+      playCount: playCount ?? this.playCount,
+      lastPlayed: lastPlayed ?? this.lastPlayed,
+      color: color ?? this.color,
     );
+  }
+
+  Voice incrementPlayCount() {
+    return copyWith(playCount: playCount + 1, lastPlayed: DateTime.now());
   }
 }
