@@ -63,15 +63,23 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
     final avatarProvider = Provider.of<AvatarProvider>(context, listen: false);
     final isDesktop = MediaQuery.of(context).size.width > 900;
     final theme = Theme.of(context);
+    final isDarkMode = theme.brightness == Brightness.dark;
 
     // Function to handle avatar creation
     void createAvatar() {
               if (_nameController.text.trim().isEmpty) {
                 // Show error message
                 ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(
-                    content: Text('Please enter a name for your avatar'),
+          SnackBar(
+            content: Text(
+              'Please enter a name for your avatar',
+              style: TextStyle(
+                color: isDarkMode ? theme.colorScheme.onInverseSurface : null,
+              ),
+            ),
                     behavior: SnackBarBehavior.floating,
+            backgroundColor:
+                isDarkMode ? theme.colorScheme.inverseSurface : null,
                   ),
                 );
                 return;
@@ -89,9 +97,296 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
       Navigator.of(context).pop();
     }
 
+    // Build the form section
+    Widget buildFormSection() {
+      return Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          // Name input with animated label
+          AnimatedContainer(
+            duration: const Duration(milliseconds: 200),
+            margin: const EdgeInsets.only(bottom: 32),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  'Avatar Name',
+                  style: TextStyle(
+                    fontSize: 16,
+                    fontWeight: FontWeight.w600,
+                    color: theme.colorScheme.onSurface,
+                  ),
+                ),
+                const SizedBox(height: 8),
+                TextField(
+                  controller: _nameController,
+                  autofocus: true,
+                  decoration: InputDecoration(
+                    hintText: 'Enter a name for your avatar',
+                    hintStyle: TextStyle(
+                      color: theme.colorScheme.onSurfaceVariant,
+                    ),
+                    border: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(12),
+                      borderSide: BorderSide(
+                        color: theme.colorScheme.outline,
+                        width: 1.5,
+                      ),
+                    ),
+                    enabledBorder: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(12),
+                      borderSide: BorderSide(
+                        color: theme.colorScheme.outline.withOpacity(0.7),
+                        width: 1.5,
+                      ),
+                    ),
+                    focusedBorder: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(12),
+                      borderSide: BorderSide(
+                        color: _getColorFromString(_selectedColor),
+                        width: 2,
+                      ),
+                    ),
+                    filled: true,
+                    fillColor:
+                        isDarkMode
+                            ? theme.colorScheme.surfaceContainerHighest
+                            : theme.colorScheme.surface,
+                    contentPadding: const EdgeInsets.symmetric(
+                      horizontal: 16,
+                      vertical: 16,
+                    ),
+                    prefixIcon: Icon(
+                      Icons.person_outline,
+                      color: theme.colorScheme.onSurfaceVariant,
+                    ),
+                  ),
+                  style: TextStyle(
+                    color: theme.colorScheme.onSurface,
+                    fontSize: 16,
+                  ),
+                  onSubmitted: (_) => createAvatar(),
+                ),
+              ],
+            ),
+          ),
+
+          // Icon selection
+          AnimatedContainer(
+            duration: const Duration(milliseconds: 200),
+            margin: const EdgeInsets.only(bottom: 32),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  'Choose an Icon',
+                  style: TextStyle(
+                    fontSize: 16,
+                    fontWeight: FontWeight.w600,
+                    color: theme.colorScheme.onSurface,
+                  ),
+                ),
+                const SizedBox(height: 16),
+                _buildIconGrid(avatarProvider.predefinedIcons),
+              ],
+            ),
+          ),
+
+          // Color selection
+          AnimatedContainer(
+            duration: const Duration(milliseconds: 200),
+            margin: const EdgeInsets.only(bottom: 32),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  'Choose a Color',
+                  style: TextStyle(
+                    fontSize: 16,
+                    fontWeight: FontWeight.w600,
+                    color: theme.colorScheme.onSurface,
+                  ),
+                ),
+                const SizedBox(height: 16),
+                _buildColorGrid(avatarProvider.predefinedColors),
+              ],
+            ),
+          ),
+
+          // Create button (for mobile)
+          if (!isDesktop)
+            SizedBox(
+              width: double.infinity,
+              child: ElevatedButton(
+                onPressed: createAvatar,
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: _getColorFromString(_selectedColor),
+                  foregroundColor: theme.colorScheme.onPrimary,
+                  padding: const EdgeInsets.symmetric(vertical: 16),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                  elevation: isDarkMode ? 4 : 2,
+                ),
+                child: const Text(
+                  'CREATE AVATAR',
+                  style: TextStyle(
+                    fontWeight: FontWeight.w600,
+                    letterSpacing: 0.5,
+                  ),
+                ),
+              ),
+            ),
+        ],
+      );
+    }
+
+    // Build the preview section
+    Widget buildPreviewSection() {
+      return AnimatedContainer(
+        duration: const Duration(milliseconds: 300),
+        curve: Curves.easeOutBack,
+        padding: const EdgeInsets.all(24),
+        decoration: BoxDecoration(
+          color:
+              isDarkMode
+                  ? theme.colorScheme.surfaceContainerHighest
+                  : theme.colorScheme.surface,
+          borderRadius: BorderRadius.circular(24),
+          border: Border.all(
+            color: _getColorFromString(_selectedColor).withOpacity(0.3),
+            width: 2,
+          ),
+          boxShadow: [
+            BoxShadow(
+              color: _getColorFromString(_selectedColor).withOpacity(0.2),
+              blurRadius: 20,
+              spreadRadius: 0,
+              offset: const Offset(0, 10),
+            ),
+          ],
+        ),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              'Preview',
+              style: TextStyle(
+                fontSize: 18,
+                fontWeight: FontWeight.w600,
+                color: theme.colorScheme.onSurface,
+                letterSpacing: 0.5,
+              ),
+            ),
+            const SizedBox(height: 32),
+            // Avatar preview with animated container
+            AnimatedContainer(
+              duration: const Duration(milliseconds: 300),
+              curve: Curves.easeOutBack,
+              width: 160,
+              height: 160,
+              decoration: BoxDecoration(
+                shape: BoxShape.circle,
+                color: _getColorFromString(_selectedColor),
+                boxShadow: [
+                  BoxShadow(
+                    color: _getColorFromString(_selectedColor).withOpacity(0.4),
+                    blurRadius: 24,
+                    spreadRadius: 4,
+                    offset: const Offset(0, 8),
+                  ),
+                ],
+              ),
+              child: Icon(_selectedIcon, size: 80, color: Colors.white),
+            ),
+            const SizedBox(height: 32),
+            // Avatar name preview
+            AnimatedDefaultTextStyle(
+              duration: const Duration(milliseconds: 200),
+              style: TextStyle(
+                fontSize: 28,
+                fontWeight: FontWeight.w600,
+                color: theme.colorScheme.onSurface,
+                letterSpacing: 0.5,
+              ),
+              child: Text(
+                _nameController.text.isEmpty
+                    ? 'Your Avatar'
+                    : _nameController.text,
+              ),
+            ),
+            const SizedBox(height: 16),
+            Text(
+              'This is how your avatar will appear in the app',
+              textAlign: TextAlign.center,
+              style: TextStyle(
+                fontSize: 14,
+                color: theme.colorScheme.onSurfaceVariant,
+              ),
+            ),
+            if (isDesktop) ...[
+              const SizedBox(height: 48),
+              // Create button for desktop in preview section
+              SizedBox(
+                width: double.infinity,
+                child: ElevatedButton(
+                  onPressed: createAvatar,
+                  style: ElevatedButton.styleFrom(
+                    backgroundColor: _getColorFromString(_selectedColor),
+                    foregroundColor: Colors.white,
+                    padding: const EdgeInsets.symmetric(vertical: 16),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12),
+                    ),
+                    elevation: 4,
+                  ),
+                  child: const Text(
+                    'CREATE AVATAR',
+                    style: TextStyle(
+                      fontWeight: FontWeight.w600,
+                      letterSpacing: 0.5,
+                      fontSize: 16,
+                    ),
+                  ),
+                ),
+              ),
+              const SizedBox(height: 16),
+              // Cancel button for desktop
+              SizedBox(
+                width: double.infinity,
+                child: OutlinedButton(
+                  onPressed: cancelCreation,
+                  style: OutlinedButton.styleFrom(
+                    foregroundColor: theme.colorScheme.onSurface,
+                    padding: const EdgeInsets.symmetric(vertical: 16),
+                    side: BorderSide(
+                      color: theme.colorScheme.outline,
+                      width: 1.5,
+                    ),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(12),
+                    ),
+                  ),
+                  child: const Text(
+                    'CANCEL',
+                    style: TextStyle(
+                      fontWeight: FontWeight.w600,
+                      letterSpacing: 0.5,
+                      fontSize: 16,
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ],
+        ),
+      );
+    }
+
     return Focus(
       autofocus: true,
-      onKeyEvent: isDesktop
+      onKeyEvent:
+          isDesktop
           ? (node, event) {
               // Handle keyboard shortcuts for desktop
               if (event.runtimeType.toString() == 'KeyDownEvent') {
@@ -101,7 +396,7 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
                   return KeyEventResult.handled;
                 }
 
-                // Enter key with Ctrl/Cmd to create
+                  // Enter key to create
                 if (event.logicalKey == LogicalKeyboardKey.enter &&
                     (HardwareKeyboard.instance.isControlPressed ||
                         HardwareKeyboard.instance.isMetaPressed)) {
@@ -113,188 +408,102 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
             }
           : null,
       child: Scaffold(
-        backgroundColor: isDesktop ? Colors.grey.shade50 : Colors.white,
+        backgroundColor: theme.colorScheme.surface,
         appBar: AppBar(
-          backgroundColor: Colors.white,
-          elevation: 0,
-          title: const Text(
+          title: Text(
             'Create New Avatar',
             style: TextStyle(
-              fontWeight: FontWeight.w400,
-              fontSize: 18,
+              fontWeight: FontWeight.w600,
+              color: theme.colorScheme.onSurface,
             ),
           ),
-          leading: IconButton(
-            icon: const Icon(Icons.close),
-            onPressed: cancelCreation,
-            tooltip: 'Cancel (Esc)',
-            style: IconButton.styleFrom(
-              foregroundColor: theme.colorScheme.onSurface,
-            ),
+          backgroundColor: Colors.transparent,
+          elevation: 0,
+          leading: _AnimatedIconButton(
+            icon: Icons.arrow_back,
+            color: theme.colorScheme.onSurface,
+            hoverColor: _getColorFromString(_selectedColor),
+            onPressed: () => Navigator.of(context).pop(),
+            tooltip: 'Back',
           ),
           actions: [
-            if (isDesktop)
-              Padding(
-                padding: const EdgeInsets.only(right: 16.0),
-                child: Tooltip(
-                  message: 'Cancel (Esc)',
-                  child: TextButton.icon(
-                    onPressed: cancelCreation,
-                    icon: const Icon(Icons.cancel_outlined, size: 18),
-                    label: const Text('Cancel'),
-                    style: TextButton.styleFrom(
-                      foregroundColor:
-                          theme.colorScheme.onSurface.withOpacity(0.8),
-                      padding: const EdgeInsets.symmetric(
-                          horizontal: 16, vertical: 8),
-                    ),
-                  ),
-                ),
-              ),
-            Padding(
-              padding: EdgeInsets.only(right: isDesktop ? 16.0 : 8.0),
-              child: isDesktop
-                  ? Tooltip(
-                      message: 'Create Avatar (Ctrl+Enter)',
-                      child: ElevatedButton.icon(
-                        onPressed: createAvatar,
-                        icon: const Icon(Icons.check, size: 18),
-                        label: const Text('Create Avatar'),
-                        style: ElevatedButton.styleFrom(
-                          foregroundColor: Colors.white,
-                          backgroundColor: theme.colorScheme.primary,
-                          padding: const EdgeInsets.symmetric(
-                              horizontal: 16, vertical: 12),
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(6),
-                          ),
-                        ),
-                      ),
-                    )
-                  : TextButton(
-                      onPressed: createAvatar,
-                      style: TextButton.styleFrom(
-                        foregroundColor: theme.colorScheme.primary,
-                        padding: const EdgeInsets.symmetric(
-                            horizontal: 16, vertical: 8),
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(6),
-                        ),
-                      ),
-                      child: const Text(
-                        'Create',
-                        style: TextStyle(
-                          fontWeight: FontWeight.w500,
-                          fontSize: 16,
-                        ),
-                      ),
-                    ),
+            if (!isDesktop)
+              _AnimatedTextButton(
+                text: 'CREATE',
+              onPressed: createAvatar,
+                color: _getColorFromString(_selectedColor),
           ),
         ],
       ),
-        body: Center(
-          child: Container(
-            width: isDesktop ? 800 : double.infinity,
-            padding: EdgeInsets.symmetric(
-              horizontal: isDesktop ? 40 : 20,
-              vertical: isDesktop ? 32 : 20,
-            ),
-            decoration: isDesktop
-                ? BoxDecoration(
-                    color: Colors.white,
-                    borderRadius: BorderRadius.circular(8),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.black.withOpacity(0.05),
-                        blurRadius: 10,
-                        spreadRadius: 0,
-                        offset: const Offset(0, 2),
-                      ),
-                    ],
-                  )
-                : null,
-            child: SingleChildScrollView(
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
+        body:
+            isDesktop
+                // Two-column layout for desktop
+                ? Row(
+                  crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
-            // Name field
-                  Column(
+                    // Form section (left column)
+                    Expanded(
+                      flex: 3,
+                      child: SingleChildScrollView(
+                        padding: const EdgeInsets.all(32.0),
+                        child: buildFormSection(),
+                      ),
+                    ),
+                    // Preview section (right column)
+                    Expanded(
+                      flex: 2,
+                      child: Container(
+                        color:
+                            isDarkMode
+                                ? theme.colorScheme.surfaceContainerLow
+                                : theme.colorScheme.surfaceContainerLowest,
+                        padding: const EdgeInsets.all(32.0),
+                        child: Center(child: buildPreviewSection()),
+                      ),
+                    ),
+                  ],
+                )
+                // Single column layout for mobile
+                : SingleChildScrollView(
+                  padding: const EdgeInsets.all(24.0),
+                  child: Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
-                      Text(
-                        'Avatar Name',
-                        style: TextStyle(
-                          fontSize: 14,
-                          fontWeight: FontWeight.w500,
-                          color: theme.colorScheme.onSurface.withOpacity(0.8),
-                        ),
-                      ),
-                      const SizedBox(height: 8),
-            TextField(
-              controller: _nameController,
-              autofocus: true,
-                        decoration: InputDecoration(
-                hintText: 'Enter avatar name',
-                          border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(8),
-                            borderSide: BorderSide(
-                              color: Colors.grey.shade300,
-                              width: 1,
-                            ),
-                          ),
-                          enabledBorder: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(8),
-                            borderSide: BorderSide(
-                              color: Colors.grey.shade300,
-                              width: 1,
-                            ),
-                          ),
-                          focusedBorder: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(8),
-                            borderSide: BorderSide(
-                              color: theme.colorScheme.primary,
-                              width: 2,
-                            ),
-                          ),
-                          contentPadding: const EdgeInsets.symmetric(
-                            horizontal: 16,
-                            vertical: 16,
-                          ),
-                          filled: true,
-                          fillColor: Colors.white,
-                        ),
-                        style: const TextStyle(
-                          fontSize: 16,
-                          fontWeight: FontWeight.w400,
-                        ),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 32),
+                      buildFormSection(),
+                      const SizedBox(height: 32),
+                      buildPreviewSection(),
+                      const SizedBox(height: 32),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
 
-            // Icon selection
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-              'Choose an Icon',
-              style: TextStyle(
-                          fontSize: 14,
-                fontWeight: FontWeight.w500,
-                          color: theme.colorScheme.onSurface.withOpacity(0.8),
-              ),
-            ),
-                      const SizedBox(height: 12),
-            Container(
+  Widget _buildIconGrid(List<IconData> icons) {
+    final avatarProvider = Provider.of<AvatarProvider>(context, listen: false);
+    final isDesktop = MediaQuery.of(context).size.width > 900;
+    final theme = Theme.of(context);
+    final isDarkMode = theme.brightness == Brightness.dark;
+
+    return Container(
                         height: 200,
               width: double.infinity,
               decoration: BoxDecoration(
-                          color: Colors.white,
-                          border: Border.all(color: Colors.grey.shade200),
-                          borderRadius: BorderRadius.circular(8),
+        color:
+            isDarkMode
+                ? theme.colorScheme.surfaceContainerHighest
+                : Colors.white,
+        border: Border.all(
+          color: isDarkMode ? theme.colorScheme.outline : Colors.grey.shade200,
+        ),
+        borderRadius: BorderRadius.circular(12),
                           boxShadow: [
                             BoxShadow(
-                              color: Colors.black.withOpacity(0.03),
+            color:
+                isDarkMode
+                    ? Colors.black.withOpacity(0.2)
+                    : Colors.black.withOpacity(0.03),
                               blurRadius: 4,
                               spreadRadius: 0,
                               offset: const Offset(0, 1),
@@ -303,16 +512,15 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
               ),
               child: GridView.builder(
                           padding: const EdgeInsets.all(16),
-                          gridDelegate:
-                              SliverGridDelegateWithFixedCrossAxisCount(
+        gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
                             crossAxisCount: isDesktop ? 8 : 5,
                             mainAxisSpacing: 16,
                             crossAxisSpacing: 16,
                   childAspectRatio: 1.0,
                 ),
-                itemCount: avatarProvider.predefinedIcons.length,
+        itemCount: icons.length,
                 itemBuilder: (context, index) {
-                  final icon = avatarProvider.predefinedIcons[index];
+          final icon = icons[index];
                   final isSelected = _selectedIcon == icon;
 
                             return Material(
@@ -329,14 +537,19 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
                       width: 50,
                       height: 50,
                       decoration: BoxDecoration(
-                        color: isSelected
-                                        ? theme.colorScheme.primary
-                                            .withOpacity(0.1)
+                  color:
+                      isSelected
+                          ? theme.colorScheme.primary.withOpacity(0.1)
+                          : isDarkMode
+                          ? theme.colorScheme.surface
                                         : Colors.grey.shade50,
                                     borderRadius: BorderRadius.circular(8),
                         border: Border.all(
-                          color: isSelected
+                    color:
+                        isSelected
                                           ? theme.colorScheme.primary
+                            : isDarkMode
+                            ? theme.colorScheme.outline
                                           : Colors.grey.shade300,
                                       width: isSelected ? 2 : 1,
                         ),
@@ -344,8 +557,11 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
                       child: Icon(
                         icon,
                         size: 28,
-                        color: isSelected
+                  color:
+                      isSelected
                                         ? theme.colorScheme.primary
+                          : isDarkMode
+                          ? theme.colorScheme.onSurface
                             : Colors.grey.shade700,
                                   ),
                       ),
@@ -353,56 +569,53 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
                   );
                 },
               ),
-            ),
-                    ],
-                  ),
-                  const SizedBox(height: 32),
+    );
+  }
 
-            // Color selection
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-              'Choose a Color',
-              style: TextStyle(
-                          fontSize: 14,
-                fontWeight: FontWeight.w500,
-                          color: theme.colorScheme.onSurface.withOpacity(0.8),
-              ),
-            ),
-                      const SizedBox(height: 12),
-            Container(
+  Widget _buildColorGrid(List<String> colors) {
+    final avatarProvider = Provider.of<AvatarProvider>(context, listen: false);
+    final isDesktop = MediaQuery.of(context).size.width > 900;
+    final theme = Theme.of(context);
+    final isDarkMode = theme.brightness == Brightness.dark;
+
+    return Container(
                         height: isDesktop ? 140 : 90,
               width: double.infinity,
               decoration: BoxDecoration(
-                          color: Colors.white,
-                          border: Border.all(color: Colors.grey.shade200),
-                          borderRadius: BorderRadius.circular(8),
+        color:
+            isDarkMode
+                ? theme.colorScheme.surfaceContainerHighest
+                : Colors.white,
+        border: Border.all(
+          color: isDarkMode ? theme.colorScheme.outline : Colors.grey.shade200,
+        ),
+        borderRadius: BorderRadius.circular(12),
                           boxShadow: [
                             BoxShadow(
-                              color: Colors.black.withOpacity(0.03),
+            color:
+                isDarkMode
+                    ? Colors.black.withOpacity(0.2)
+                    : Colors.black.withOpacity(0.03),
                               blurRadius: 4,
                               spreadRadius: 0,
                               offset: const Offset(0, 1),
                             ),
                           ],
                         ),
-                        child: isDesktop
+      child:
+          isDesktop
                             // Grid layout for desktop
                             ? GridView.builder(
                                 padding: const EdgeInsets.all(16),
-                                gridDelegate:
-                                    const SliverGridDelegateWithFixedCrossAxisCount(
+                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
                                   crossAxisCount: 10,
                                   mainAxisSpacing: 16,
                                   crossAxisSpacing: 16,
                                   childAspectRatio: 1.0,
                                 ),
-                                itemCount:
-                                    avatarProvider.predefinedColors.length,
+                itemCount: colors.length,
                                 itemBuilder: (context, index) {
-                                  final color =
-                                      avatarProvider.predefinedColors[index];
+                  final color = colors[index];
                                   final isSelected = _selectedColor == color;
 
                                   return InkWell(
@@ -413,28 +626,26 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
                                     },
                                     borderRadius: BorderRadius.circular(30),
                                     child: AnimatedContainer(
-                                      duration:
-                                          const Duration(milliseconds: 200),
+                      duration: const Duration(milliseconds: 200),
                                       decoration: BoxDecoration(
                                         color: _getColorFromString(color),
                                         shape: BoxShape.circle,
                                         border: Border.all(
-                                          color: isSelected
-                                              ? Colors.white
-                                              : Colors.transparent,
+                          color: isSelected ? Colors.white : Colors.transparent,
                                           width: 3,
                                         ),
                                         boxShadow: [
                                           BoxShadow(
-                                            color: _getColorFromString(color)
-                                                .withOpacity(
-                                                    isSelected ? 0.5 : 0.3),
+                            color: _getColorFromString(
+                              color,
+                            ).withOpacity(isSelected ? 0.5 : 0.3),
                                             blurRadius: isSelected ? 12 : 4,
                                             spreadRadius: isSelected ? 2 : 0,
                                           ),
                                         ],
                                       ),
-                                      child: isSelected
+                      child:
+                          isSelected
                                           ? const Icon(
                                               Icons.check,
                                               color: Colors.white,
@@ -449,11 +660,9 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
                             : ListView.builder(
                 scrollDirection: Axis.horizontal,
                                 padding: const EdgeInsets.all(16),
-                                itemCount:
-                                    avatarProvider.predefinedColors.length,
+                itemCount: colors.length,
                 itemBuilder: (context, index) {
-                                  final color =
-                                      avatarProvider.predefinedColors[index];
+                  final color = colors[index];
                   final isSelected = _selectedColor == color;
 
                   return Padding(
@@ -466,30 +675,29 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
                       },
                                       borderRadius: BorderRadius.circular(30),
                                       child: AnimatedContainer(
-                                        duration:
-                                            const Duration(milliseconds: 200),
+                        duration: const Duration(milliseconds: 200),
                                         width: 56,
                                         height: 56,
                         decoration: BoxDecoration(
                           color: _getColorFromString(color),
                           shape: BoxShape.circle,
                           border: Border.all(
-                                            color: isSelected
-                                                ? Colors.white
-                                                : Colors.transparent,
+                            color:
+                                isSelected ? Colors.white : Colors.transparent,
                                             width: 3,
                                           ),
                                           boxShadow: [
                                   BoxShadow(
-                                    color: _getColorFromString(color)
-                                                  .withOpacity(
-                                                      isSelected ? 0.5 : 0.3),
+                              color: _getColorFromString(
+                                color,
+                              ).withOpacity(isSelected ? 0.5 : 0.3),
                                               blurRadius: isSelected ? 12 : 4,
                                               spreadRadius: isSelected ? 2 : 0,
                                             ),
                                           ],
                                         ),
-                                        child: isSelected
+                        child:
+                            isSelected
                                             ? const Icon(
                                                 Icons.check,
                                                 color: Colors.white,
@@ -501,95 +709,135 @@ class _AddAvatarScreenState extends State<AddAvatarScreen> {
                   );
                 },
               ),
-                      ),
-                    ],
-            ),
+    );
+  }
+}
 
-            // Preview section
-            const SizedBox(height: 40),
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-              'Preview',
-              style: TextStyle(
-                          fontSize: 14,
-                fontWeight: FontWeight.w500,
-                          color: theme.colorScheme.onSurface.withOpacity(0.8),
+// Animated Icon Button for app bar icons
+class _AnimatedIconButton extends StatefulWidget {
+  final IconData icon;
+  final Color color;
+  final Color hoverColor;
+  final VoidCallback onPressed;
+  final String tooltip;
+
+  const _AnimatedIconButton({
+    required this.icon,
+    required this.color,
+    required this.hoverColor,
+    required this.onPressed,
+    required this.tooltip,
+  });
+
+  @override
+  State<_AnimatedIconButton> createState() => _AnimatedIconButtonState();
+}
+
+class _AnimatedIconButtonState extends State<_AnimatedIconButton> {
+  bool _isHovering = false;
+
+  @override
+  Widget build(BuildContext context) {
+    // Only use hover effect on desktop/web platforms
+    final bool isMobile = MediaQuery.of(context).size.width < 600;
+
+    return MouseRegion(
+      onEnter: (_) => isMobile ? null : setState(() => _isHovering = true),
+      onExit: (_) => isMobile ? null : setState(() => _isHovering = false),
+      cursor: SystemMouseCursors.click,
+      child: Tooltip(
+        message: widget.tooltip,
+        child: IconButton(
+          icon: TweenAnimationBuilder<double>(
+            tween: Tween<double>(begin: 0, end: _isHovering ? 1.0 : 0.0),
+            duration: const Duration(milliseconds: 200),
+            curve: Curves.easeOutBack,
+            builder: (context, value, child) {
+              return Transform.scale(
+                scale: 1.0 + (value * 0.2),
+                child: AnimatedContainer(
+                  duration: const Duration(milliseconds: 200),
+                          decoration: BoxDecoration(
+          color:
+                        _isHovering
+                            ? widget.hoverColor.withOpacity(0.1)
+                            : Colors.transparent,
+                    shape: BoxShape.circle,
+                  ),
+                  padding: const EdgeInsets.all(4),
+                  child: Icon(
+                    widget.icon,
+                    color: Color.lerp(widget.color, widget.hoverColor, value),
+                    size: 24,
+                  ),
+                ),
+              );
+            },
+          ),
+          onPressed: widget.onPressed,
+          splashRadius: 24,
+        ),
+      ),
+    );
+  }
+}
+
+// Animated Text Button for app bar actions
+class _AnimatedTextButton extends StatefulWidget {
+  final String text;
+  final VoidCallback onPressed;
+  final Color color;
+
+  const _AnimatedTextButton({
+    required this.text,
+    required this.onPressed,
+    required this.color,
+  });
+
+  @override
+  State<_AnimatedTextButton> createState() => _AnimatedTextButtonState();
+}
+
+class _AnimatedTextButtonState extends State<_AnimatedTextButton> {
+  bool _isHovering = false;
+
+  @override
+  Widget build(BuildContext context) {
+    // Only use hover effect on desktop/web platforms
+    final bool isMobile = MediaQuery.of(context).size.width < 600;
+
+    return MouseRegion(
+      onEnter: (_) => isMobile ? null : setState(() => _isHovering = true),
+      onExit: (_) => isMobile ? null : setState(() => _isHovering = false),
+      cursor: SystemMouseCursors.click,
+      child: TweenAnimationBuilder<double>(
+        tween: Tween<double>(begin: 0, end: _isHovering ? 1.0 : 0.0),
+        duration: const Duration(milliseconds: 200),
+        curve: Curves.easeOutBack,
+        builder: (context, value, child) {
+          return TextButton(
+            onPressed: widget.onPressed,
+            style: TextButton.styleFrom(
+              foregroundColor: widget.color,
+              backgroundColor: widget.color.withOpacity(value * 0.1),
+              padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(8),
               ),
             ),
-                      const SizedBox(height: 24),
-            Center(
-              child: Container(
-                          padding: const EdgeInsets.all(32),
-                          decoration: BoxDecoration(
-                            color: Colors.white,
-                            borderRadius: BorderRadius.circular(16),
-                            border: Border.all(color: Colors.grey.shade200),
-                            boxShadow: [
-                              BoxShadow(
-                                color: Colors.black.withOpacity(0.05),
-                                blurRadius: 10,
-                                spreadRadius: 0,
-                                offset: const Offset(0, 4),
-                              ),
-                            ],
-                          ),
-                          child: Column(
-                            children: [
-                              Container(
-                width: 120,
-                height: 120,
-                decoration: BoxDecoration(
-                  shape: BoxShape.circle,
-                  color: _getColorFromString(_selectedColor),
-                  boxShadow: [
-                    BoxShadow(
-                                      color: _getColorFromString(_selectedColor)
-                                          .withOpacity(0.3),
-                                      blurRadius: 16,
-                      spreadRadius: 2,
-                                      offset: const Offset(0, 6),
-                    ),
-                  ],
-                ),
-                child: Icon(
-                  _selectedIcon,
-                  size: 60,
-                  color: Colors.white,
-                ),
-              ),
-                              const SizedBox(height: 20),
-                              Text(
-                _nameController.text.isEmpty
-                    ? 'Your Avatar'
-                    : _nameController.text,
-                style: const TextStyle(
-                                  fontSize: 24,
-                                  fontWeight: FontWeight.w400,
-                                ),
-                              ),
-                              if (isDesktop) ...[
-                                const SizedBox(height: 24),
-                                Text(
-                                  'This is how your avatar will appear in the app',
+            child: Transform.scale(
+              scale: 1.0 + (value * 0.05),
+              child: Text(
+                widget.text,
                                   style: TextStyle(
-                                    fontSize: 14,
-                                    color: Colors.grey.shade600,
+                  fontWeight: FontWeight.w600,
+                  letterSpacing: 0.5,
+                  color: widget.color,
                                   ),
                                 ),
-                              ],
-                            ],
-                ),
-              ),
-            ),
-          ],
-                  ),
-                ],
-              ),
-            ),
-          ),
         ),
+          );
+        },
       ),
     );
   }
