@@ -24,54 +24,77 @@ class _VoicePlayerState extends State<VoicePlayer> {
   double _progress = 0.0;
   StreamSubscription? _positionSubscription;
   StreamSubscription? _playerStateSubscription;
+  bool _isInitialized = false;
 
   @override
   void initState() {
     super.initState();
-    _initializePlayer();
+    // Use a microtask to ensure initialization happens after the widget is built
+    // This helps prevent threading issues
+    Future.microtask(() => _initializePlayer());
   }
 
   @override
   void dispose() {
-    _positionSubscription?.cancel();
-    _playerStateSubscription?.cancel();
-    _player.dispose();
+    // Cancel subscriptions first to prevent callbacks after disposal
+    _cancelSubscriptions();
+    // Use a try-catch to handle any errors during disposal
+    try {
+      _player.dispose();
+    } catch (e) {
+      debugPrint('Error disposing player: $e');
+    }
     super.dispose();
+  }
+
+  void _cancelSubscriptions() async {
+    try {
+      await _positionSubscription?.cancel();
+      _positionSubscription = null;
+      await _playerStateSubscription?.cancel();
+      _playerStateSubscription = null;
+    } catch (e) {
+      debugPrint('Error cancelling subscriptions: $e');
+    }
   }
 
   @override
   void didUpdateWidget(VoicePlayer oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (oldWidget.audioPath != widget.audioPath) {
-      _initializePlayer();
+      // Use a microtask to ensure initialization happens after the widget is updated
+      Future.microtask(() => _initializePlayer());
     }
   }
 
   Future<void> _initializePlayer() async {
-    try {
-      // Cancel existing subscriptions
-      await _positionSubscription?.cancel();
-      await _playerStateSubscription?.cancel();
+    // Cancel existing subscriptions to prevent memory leaks
+    _cancelSubscriptions();
 
+    try {
       // Set up the audio source
       await _player.setSource(DeviceFileSource(widget.audioPath));
 
-      // Set up position stream
+      // Set up position stream with error handling
       _positionSubscription = _player.onPositionChanged.listen(
         (position) {
           if (mounted) {
             setState(() {
               _currentPosition = position;
-              _progress = position.inMilliseconds /
+              _progress =
+                  position.inMilliseconds /
                   (widget.duration.inMilliseconds == 0
                       ? 1
                       : widget.duration.inMilliseconds);
             });
           }
         },
+        onError: (error) {
+          debugPrint('Position stream error: $error');
+        },
       );
 
-      // Set up player state stream
+      // Set up player state stream with error handling
       _playerStateSubscription = _player.onPlayerStateChanged.listen(
         (state) {
           if (mounted) {
@@ -84,13 +107,34 @@ class _VoicePlayerState extends State<VoicePlayer> {
             });
           }
         },
+        onError: (error) {
+          debugPrint('Player state stream error: $error');
+        },
       );
+
+      // Mark as initialized
+      if (mounted) {
+        setState(() {
+          _isInitialized = true;
+        });
+      }
     } catch (e) {
       debugPrint('Error initializing player: $e');
+      if (mounted) {
+        setState(() {
+          _isInitialized = false;
+        });
+      }
     }
   }
 
   Future<void> _togglePlayPause() async {
+    if (!_isInitialized) {
+      // Try to initialize again if not initialized
+      await _initializePlayer();
+      if (!_isInitialized) return;
+    }
+
     try {
       if (_isPlaying) {
         await _player.pause();
@@ -106,6 +150,8 @@ class _VoicePlayerState extends State<VoicePlayer> {
   }
 
   Future<void> _seekTo(double value) async {
+    if (!_isInitialized) return;
+
     try {
       final position = Duration(
         milliseconds: (widget.duration.inMilliseconds * value).round(),
@@ -151,17 +197,11 @@ class _VoicePlayerState extends State<VoicePlayer> {
               children: [
                 Text(
                   _formatDuration(_currentPosition),
-                  style: TextStyle(
-                    fontSize: 12,
-                    color: Colors.grey[600],
-                  ),
+                  style: TextStyle(fontSize: 12, color: Colors.grey[600]),
                 ),
                 Text(
                   _formatDuration(widget.duration),
-                  style: TextStyle(
-                    fontSize: 12,
-                    color: Colors.grey[600],
-                  ),
+                  style: TextStyle(fontSize: 12, color: Colors.grey[600]),
                 ),
               ],
             ),
@@ -182,9 +222,7 @@ class _VoicePlayerState extends State<VoicePlayer> {
 class _AudioWaveform extends StatefulWidget {
   final bool isPlaying;
 
-  const _AudioWaveform({
-    required this.isPlaying,
-  });
+  const _AudioWaveform({required this.isPlaying});
 
   @override
   State<_AudioWaveform> createState() => _AudioWaveformState();
@@ -225,11 +263,7 @@ class _AudioWaveformState extends State<_AudioWaveform>
   Widget build(BuildContext context) {
     return Row(
       children: [
-        Icon(
-          Icons.volume_up,
-          size: 16,
-          color: Colors.grey[600],
-        ),
+        Icon(Icons.volume_up, size: 16, color: Colors.grey[600]),
         const SizedBox(width: 4),
         SizedBox(
           width: 16,
@@ -241,12 +275,16 @@ class _AudioWaveformState extends State<_AudioWaveform>
               (index) => AnimatedBuilder(
                 animation: _controller,
                 builder: (context, child) {
-                  final value = widget.isPlaying
-                      ? (1.0 +
-                              sin((_controller.value * 3.14 + index * 0.5)) *
-                                  0.3)
-                          .clamp(0.3, 1.0)
-                      : 0.3;
+                  final value =
+                      widget.isPlaying
+                          ? (1.0 +
+                                  sin(
+                                        (_controller.value * 3.14 +
+                                            index * 0.5),
+                                      ) *
+                                      0.3)
+                              .clamp(0.3, 1.0)
+                          : 0.3;
                   return Container(
                     width: 2,
                     height: 12 * value,
